Refactor: Move redirection logic to a separate class
